Increase bot's movement speed
Now that there are other vehicles (Shotgun Tanks!) the role of Bots
needs to change. Their advantage is becoming speed: fast enough to
evade fire when on the periphery of turrets. 3.5 bots can kill a
Shotgun Tank; making them faster is needed for them to be able to
actually do it.