Add prototype superweapon-style Titan unit
The Titan isn't buildable yet (except via the console) but it offers
frightening levels of firepower. Against a mass of turrets it isn't
strong enough to survive, but it can take down a few turrets and it
has enough firepower to cripple a mass of energy units. It's sort-of
a longer-range super-shotgun-tank.