Allow human orders for AttackNearestAI vehicles
Before this commit, AttackNearestAI would constantly change the orders
of all its vehicles and so human orders would get discarded. This
commit starts keeping a record of the orders that have been assigned
to non-Commander vehicles, and detects when a human changed an order
so it shouldn't change the order.