Make shotgun tanks rotate to fire projectiles
This commit includes a few nasty hacks. It delivers a change to the
behaviour of Shotgun Tanks. Until now Shotgun Tanks have been able to
fire in any direction at any time, just like all other units. This is
very powerful–too powerful–and it takes away a lot of tactical depth.

With this commit in place, Shotgun Tanks have to turn to face their
targets. The firing direction is completely independent of the vehicle
direction, although if there are no enemy units nearby the unit will
slowly rotate to be looking straight ahead. When an enemy is nearby,
even if out of range, it rotates to be ready to fire at the nearest
enemy. And when an enemy is almost within firing range it opens fire,
even if it is not pointing in the right direction.