Make titans rotate to fire projectiles
This commit duplicates the previous commit's work on Shotgun Tanks
but applies it to Titans. I want to make the game require a bit more
micro and provide more reward for thinking through tactics. Making
units only able to fire in particular directions opens the
opportunity for outmaneuvering your enemy.