Support maps with unpassable "obstruction" squares
This commit adds support and rendering for Obstructions. These are
square areas that units cannot cross, but projectiles can. This is the
very first movement restriction in the game and it's very improvised,
but it's a good simple start.

This gets RTT one step closer to real Real-Time Strategy games. A lot
of maps in RTS games have some obstacles or unpassable areas in the
map, which makes the game more interesting and challenging.