Recalculate paths regularly to improve movement
There are three systems involved in vehicles moving around:

  1. The point-to-point routes obtained from `nav2d`, giving a series
     of waypoints from where the vehicle started to where the vehicle
     wants to go

  2. The vehicle steering making it turn towards the next waypoint on
     its route

  3. The collision detection preventing vehicles from crossing or
     going inside obstacles

All three of these systems in RTT are quite improvised. I haven't done
any of these on this scale before.

Steering vehicles in RTT have a minimum turning circle. That is they
can't turn on a dime. This is similar to real world cars.

The steering only tries to turn towards the next waypoint. If there is
an obstacle in the way of the turning circle needed to reach the next
waypoint, the vehicle can get stuck because it hits the obstacle and
stops moving productively.

This commit has vehicles regularly update the route from their current
position to their destination. This seems to prevent them getting
stuck, by ensuring that units are nudged in the right direction when
colliding with obstacles.

It's probably best to solve this problem in a more integrated way:
consider the vehicle turning circle as part of the point-to-point
routes. Unfortunately I don't even know where to get started on that
without making pathfinding far more expensive.

This commit will impact performance, but for now it's the best idea
that I've got.