Fix combination of pathfinding/collision detection
Before this commit there was only one navmesh that unit paths were
calculated on. The navmesh didn't include triangles for the obstacles,
just holes. The navmesh didn't quite touch the obstacles, so that the
calculated paths reduced the number of collisions.

A problem with that setup is that a unit colliding with an obstacle
could end up too close to the obstacle and off of the navmesh. If the
path was recomputed then the unit was stuck, because it was off the
navmesh and therefore had no path.

This problem was enhanced by when other units attempted to approach
that unit to attack. Those units became stuck too, because they could
not path to the unit!

This commit fixes that situation by creating a second navmesh. This
one still has holes for the obstacles, but it touches the obstacles.
This means that the unit can still find a path so long as the unit
has been pushed out of the collision first. This new navmesh is used
as a backup when no path is found.

This isn't quite good enough for the long-term, but it's a lot more
solid than the collision detection and seems to work reliably.