[WIP] [#14] Experiment with factory colour animation
In https://github.com/46bit/rts, factories had a solid border colour
with a varying inside area. The inside area changed colour to
visualise the progress of constructing a new unit.

This commit experiments with replicating it in WebGL. A few issues:

  - Right now I use a perspective camera, not orthographic. This is
    so that I can reuse CameraControls, a nice library for users to
    move the camera around. But the perspective is visible when
    meshes are being visually layered at different layers from the
    camera.

  - This really shouldn't be done using multiple meshes. It should be
    done using a texture and then a custom shader. But I don't want
    to open that pandora's box quite yet.

So I'll leave this for later. Saving this work as a record of how to
do custom shaders and notes above.