Switch to an Orthographic camera
Until now I've been a bit lazy and used a perspective camera. The
perspective distortion was small enough to be unnoticable in what is
still a 2D game. But as I move into more complex visual effects it has
become a problem.

This commit switches to an orthographic camera. For the time being it
also removes the ability to zoom in/out and move the camera around.
That isn't essential and I can reintroduce it as part of improving the
user interaction. :)