Render entity selection in better code
This removes a lot more logic from the `main` function, and reproduces
it in decent code elsewhere. After this commit we're completely set
for selecting units. All that's left is to limit the selection to
units of a specific player--and I'm not there with everything else
needed for a true single-player mode yet.