Fix obstruction circumcircle radius
I use a quadtree to speed up checking for collisions. The quadtree
gives a performance of O(nlog(n))-ish. This is much better than the
O(n^2) of having to check every single entity against each other.

One of the many types of collisions I'm looking for is where a unit
has collided with an Obstruction. Obstructions are rectangular, but
my quadtree only deals with circles. So I use the quadtree to detect
possible collisions, and then check those to see if there was an
actual collision.

I've been investigating a failure of the collision code, and I
discovered that I was looking for possible collisions wrongly. The
circumcircle wasn't big enough to encompass the corners of the
obstruction. I hadn't noticed this before because until now I've used
obstacles where the corners were a small part of their surface area.
Now I'm using lots of small obstacles this has caused more of a
problem. More info at [1].

This commit fixes the obstruction circumcircle code so that it
includes the corners of the circumcircle.

[1] https://github.com/46bit/rtt/issues/15#issuecomment-685165203