Zoom in and out by scrolling the mouse
When the user scrolls in, they will zoom into the area around their
cursor. They can't scroll in closer than 250px. They also can't scroll
out further than the point at which the entire board is visible, and
there is logic to ensure zooming out doesn't go off the board.

This commit makes several long-needed tweaks to how positioning of the
camera, scene, etc is done. For now that global positioning is
configured on `camera.matrix` so that it is available to
`THREE.Raytracer`.