Add engineers & support group builds of power gens
This commit adds a new type of unit, the Engineer. It is slower than
the Bot but has a little more health and is a little less expensive.
The main use for an Engineer is to build new structures. The engineer
builds a bit slower (~70%) of the speed at which a Commander builds.

When another unit collides with an Engineer, the Engineer deals a lot
less damage than a typical collision (roughly 1/6th of the Engineer's
remaining health.) This is to prevent Engineer spam from replacing
Bots as an offensive option. The damage is not zero to prevent a
single Bot from moving down vast numbers of engineers.

This commit also adds support for multiple Engineers to build a single
power generator. Basically the code supports discovering that a
power generator at that location is already under construction.

I chose that Engineers with a power gen build order, where the power
gen is completed after they've been issued with the order, won't
cancel the order until they get within build range. The player can
manage that and it will prevent engineers just… stopping, in the
middle of nowhere.

As previously noted, the entire construction and ordering system is a
mess. At some stage it needs a major revisit. Some of it doesn't even
typecheck. But it does work alright for now.

Closes #25. Makes #7 a bit more urgent.