Support a rectangular, resizable viewport
Before this commit the render canvas was hardcoded to be square. That
made for a quicker implementation.

This commit changes the canvas to fill up the entire screen space,
except for the area taken up by the sidebar. The zooming and movement
logic turned out to still work perfectly.

This commit also makes the canvas resize on the next animation frame
after the browser is resized.