Avoid circular reference between dead units
It's pretty common for two enemy units to be attacking each other when
they die. Until now that could result in a memory leak. If both units
were ordered to attack each other when they died, they would have
circular references [1] when disposed of.

Preventing this memory leak should be as easy as removing all the
orders a unit had when it dies.

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Memory_Management#Limitation_Circular_references