Improve how ExpansionAI builds power generators
Since ExpansionAI exists to collect an overwhelming amount of energy,
it is important that it expands as fast as possible. It is only a demo
AI, but it should still be pretty good.

Until now, ExpansionAI's engineers have naively picked the closest
unoccupied or unfinished power source and gone to build it. This led
to lots of engineers picking the same empty power sources, leaving
others slightly further away unoccupied.

This commit puts in place a new approach to choose where to send
engineers. Engineers will still build on the closest power source,
EXCEPT if another engineer is already headed to build there. If other
engineers are already headed there, the distance is multiplied by the
number of those engineers.

This encourages engineers to choose power sources up to a few times as
far away rather than mob the same single power source.

This only applies if the AI has more than 3 power generators built,
ensuring the early game expansion isn't too slow.