Enable assigning attack orders
If the player selects attack and then right clicks on a unit, attack
orders are issued. Even if the same player owns that unit.

If there's no unit where the right-click is, the unit is ordered to
go there instead.

This isn't a long-term solution, but it's progress for the time being.