Stop using classes for entity abilities
This commit makes an experimental, forward-looking change. Until now
each entity (e.g., a factory or a shotgun tank) has been an instance
of a class. Classes such as `Factory` or `ShotgunTank` were assembled
using a custom class mixin/multiple-inheritance setup.

I built that to see if I could rather than really expecting it to
last. But it worked so well that it lasted until now! This commit
removes the entity classes, in favour of plain interfaces with
functions acting on them.

This is not the final word in how this will work, but it's the start
of a significant departure. I am hopeful that this will make tasks
such as serialisation, recording state and testing much easier.