Switch pathfinding library from nav2d to navmesh
A project on Github benchmarked a few browser navmesh libraries and
decided that this navmesh library [1] was the fastest. It estimated it to
be about 30x faster than the nav2d library [2] I've been using up to now
[3].

In this commit I adopt that navmesh library. Running the game and
looking at the timings in `console.log`, I can see a vast improvement.
In the late game (400+ units) nav2d could eat 70% of the compute time.
With navmesh, pathfinding rarely even appears in the list of things
taking up compute time.

This is particularly interesting to me given that nav2d seems to have
some performance enhancements that navmesh does not. For instance,
nav2d finds the polygon containing the start/end points using a
quadtree. Navmesh almost brute-forces it [4].

Navmesh doesn't seem to do anything particularly special compared to
nav2d. If anything, nav2d looks more optimised. I suspect Navmesh is
only faster because of the third-party A* library [5] that it uses for
pathfinding across the navmesh.

[1] https://github.com/mikewesthad/navmesh
[2] https://github.com/frapa/nav2d
[3] https://github.com/will-hart/2d-navmesh-experiment/commit/31ae6bca7f09822bb5bd2280274bd18b18345d7b
[4] https://github.com/mikewesthad/navmesh/blob/f0dde54703bbc8a69592196c7aa2c1d543f1ea62/packages/navmesh/src/navmesh.js#L86
[5] https://github.com/mikewesthad/navmesh/blob/master/packages/navmesh/package.json#L14