Reduce rate of rendering allocations
This commit alters the logic of the main way of drawing entities. We
use "instanced geometries" to efficiently render many entities of the
same type: we send the GPU one copy of the entity geometry, and then
send instance parameters to render that many times.

Until now we have reinitialised the instancing every time the number
of instances of a given type changed. That was inefficient, as we can
cheaply render fewer items than are allocated.

After this commit, we only reinitialise when we do not have enough
memory allocated to represent the number of instances. So if the
number of instances needed drops we don't do any allocations.

Also in this commit, we grow the instanced buffers by 32 entities at
a time. This will reduce the number of allocations that happen as new
entities are created.

The memory used to store extra instances is not going to be a big deal
most of the time. However, it may make sense to start freeing truly
unneeded memory. If we shard instances across several buffers it
would allow for partially freeing memory. However I don't think that
is worthwhile: the extra draw calls will destroy performance. What
would be better future work is to reinitialise the buffers if we are
consistently a long way below what is allocated.