Force a half-second delay between building things
Until now there has been an easy way to confuse the AIs: build a
building closer to their units than you are, and keep rebuilding it
every time it gets killed. Since you could start the next building
on the next tick after the previous one died, your opponent might
always attack that building and never attack you.

This commit fixes that problem for the time being. The engineer code
is embarrasingly bad, but this fixes the problem until I go refactor
it. There is now a half-second (30 tick) period before you can start
the next building. This is long enough for enemies to have retargeted
and attacked you–particularly when repeated.

Closes #7.